requirements.txt
flask
requests
paho-mqtt
psutil
